fix(default.nix): 🐛 Update hostname for live USB configuration
Changed the hostname in the live USB configuration from "vm" to "live-usb" for better clarity and consistency. This update ensures that the configuration accurately reflects the intended use case.

Additionally, added a section in the README for building the ISO image, providing users with clear instructions on how to create a bootable USB image from the NixOS configuration.