update(flake): ⬆️ Update locked dependencies
Updated the locked dependencies in the flake.lock file to reflect the latest versions. This includes changes to the last modified timestamps and hash values for various dependencies, ensuring that the project uses the most recent and secure versions available.

- Updated lastModified and narHash for multiple dependencies.
- Updated revision hashes for autofirma-nix, firefox-addons, home-manager, and nixpkgs.

This change helps maintain the integrity and security of the project dependencies.