add(cli): ➕ Add new packages to CLI configuration
Added 'python313Packages.aubio', 'yt-dlp', 'p7zip', and '_7zz' to the CLI package list for enhanced functionality. This update expands the available tools for users, improving the overall utility of the CLI environment.

Also included 'gearlever' and 'geary' in the GTK package list to provide additional applications for user convenience.