remove(readme): ➖ Remove outdated usage instructions
Deleted the usage section from the README.md file to streamline the documentation. The removed instructions included commands for building ISO images, installing the OS and home configurations, updating flakes, setting secrets, and retrieving cached revisions. This change aims to reduce clutter and improve the overall clarity of the README.