update(flake.lock): ⬆️ Update nix-chuwi-minibook-x lock file
Updated the lock file for nix-chuwi-minibook-x to reflect the latest changes. This includes modifications to the last modified timestamp, narHash, and revision number to ensure compatibility with the latest dependencies.

This update is necessary to maintain the integrity and functionality of the project with the latest changes in the dependencies.