update(flake.lock): ⬆️ Update locked dependencies
Updated the locked dependencies in flake.lock to reflect the latest versions. This includes changes to the last modified timestamps and nar hashes for various packages, ensuring compatibility with the latest updates.

Also added a new overlay for niri in overlays/niri.nix to override attributes for the niri package, disabling the doCheck option.