update(flake): ⬆️ Update flake.lock and package configurations
Updated the flake.lock file to reflect the latest versions of dependencies. This includes changes to the last modified timestamps and hashes for several packages. Additionally, made adjustments to the cli.nix and gtk.nix files to comment out unused packages and add new ones, enhancing the overall package management.

- Updated lastModified and narHash for multiple packages in flake.lock.
- Commented out python313Packages.markitdown in cli.nix.
- Added wf-recorder to cli.nix.
- Reorganized and added several packages in gtk.nix.