update(flake): ⬆️ Update flake.lock and flake.nix configurations
Updated the flake.lock file to reflect the latest changes in dependencies, including modifications to the lastModified timestamp and narHash for the home-manager repository.

Additionally, cleaned up the flake.nix file by removing unnecessary blank lines and commenting out the postInstall script for the code package, which may be revisited in the future.

This ensures that the project is using the most recent configurations and maintains a cleaner codebase.